+++ test/test_helper.rb
@@ -109,9 +109,10 @@ module KubernetesDeploy
       end
 
       logging_assertion do |logs|
-        assert_match(/failed due to timeouts./, logs) if cause == :timeout
+        cause_string = cause == :timed_out ? "TIMED OUT" : "FAILURE"
+        assert_match Regexp.new("Result: #{cause_string}"), logs,
+          "'Result: #{cause_string}' not found in the following logs:\n#{logs}"
         assert_equal false, result, "Deploy succeeded when it was expected to fail. Logs:\n#{logs}"
-        assert_match Regexp.new("Result: FAILURE"), logs, "'Result: FAILURE' not found in the following logs:\n#{logs}"
       end
     end
     alias_method :assert_restart_failure, :assert_deploy_failure