+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### 0.17.1
 *Features*
-- kubernetes-restart and kubernetes-deploy use exit code 124 when a
+- kubernetes-restart and kubernetes-deploy use exit code 70 when a
 deploy fails due to one or more resources failing to deploy in time.
 ([#244](https://github.com/Shopify/kubernetes-deploy/pull/244))